add new type and relocate
