+++ src/atoms/AllDocumentsData.ts
@@ -1,0 +1,6 @@
+import { atom } from "recoil";
+
+export const allDocumentsState = atom<{ [key: string]: string }>({
+  key: "allDocumentsState",
+  default: {},
+});
