+++ src/types/titleSetsType.ts
@@ -1,0 +1,1 @@
+export type titleSetsType = { [key: string]: string };
